requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.23.4
google-api-python-client==2.108.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1
pandas==2.1.3
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
